refactor: move the research Resource catalog from sharing-service into research-service
ResourceEntity (table RESOURCE), TagEntity, ResourceStarEntity, their three repositories and the four research-only enums (ResourceTypeEnum/PrivacyEnum/StateEnum/StatusEnum) physically lived in sharing-service but were referenced only by research-service, which reached into sharing's internal model/repository packages 28 times. This is a package-only relocation (org.apache.airavata.sharing.{model,repository} -> research.{model,repository}) with no schema change: the RESOURCE/TAG/RESOURCE_STAR tables and their Flyway migrations are untouched, and the broad @EntityScan over org.apache.airavata still discovers them. All consumers and the shared test persistence unit are repointed; research-service's sharing.model/repository imports drop from 28 to 0. Behavior-preserving; the reactor builds green and research-service's 53 tests pass.
