+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/config/DevDataInitializer.java
@@ -22,14 +22,14 @@ package org.apache.airavata.research.config;
 import java.util.HashSet;
 import java.util.Set;
 import org.apache.airavata.research.model.DatasetResourceEntity;
+import org.apache.airavata.research.model.PrivacyEnum;
 import org.apache.airavata.research.model.RepositoryResourceEntity;
 import org.apache.airavata.research.model.ResearchProjectEntity;
+import org.apache.airavata.research.model.StatusEnum;
+import org.apache.airavata.research.model.TagEntity;
 import org.apache.airavata.research.repository.ResearchProjectRepository;
-import org.apache.airavata.sharing.model.PrivacyEnum;
-import org.apache.airavata.sharing.model.StatusEnum;
-import org.apache.airavata.sharing.model.TagEntity;
-import org.apache.airavata.sharing.repository.ResourceRepository;
-import org.apache.airavata.sharing.repository.TagRepository;
+import org.apache.airavata.research.repository.ResourceRepository;
+import org.apache.airavata.research.repository.TagRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
