+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/model/PrivacyEnum.java
@@ -17,7 +17,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.apache.airavata.sharing.model;
+package org.apache.airavata.research.model;
 
 public enum PrivacyEnum {
     PUBLIC,
