+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/model/ResearchProjectEntity.java
@@ -35,7 +35,6 @@ import jakarta.persistence.Table;
 import java.time.Instant;
 import java.util.HashSet;
 import java.util.Set;
-import org.apache.airavata.sharing.model.StateEnum;
 import org.hibernate.annotations.UuidGenerator;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
