+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/model/StateEnum.java
@@ -17,11 +17,9 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.apache.airavata.sharing.model;
+package org.apache.airavata.research.model;
 
-public enum StatusEnum {
-    NONE,
-    PENDING,
-    VERIFIED,
-    REJECTED,
+public enum StateEnum {
+    ACTIVE,
+    DELETED
 }
