+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/model/StatusEnum.java
@@ -17,9 +17,11 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.apache.airavata.sharing.model;
+package org.apache.airavata.research.model;
 
-public enum StateEnum {
-    ACTIVE,
-    DELETED
+public enum StatusEnum {
+    NONE,
+    PENDING,
+    VERIFIED,
+    REJECTED,
 }
