+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/service/ResearchProjectService.java
@@ -29,11 +29,11 @@ import org.apache.airavata.config.UserContext;
 import org.apache.airavata.research.model.DatasetResourceEntity;
 import org.apache.airavata.research.model.RepositoryResourceEntity;
 import org.apache.airavata.research.model.ResearchProjectEntity;
+import org.apache.airavata.research.model.ResourceEntity;
+import org.apache.airavata.research.model.ResourceTypeEnum;
+import org.apache.airavata.research.model.StateEnum;
 import org.apache.airavata.research.repository.ResearchProjectRepository;
-import org.apache.airavata.sharing.model.ResourceEntity;
-import org.apache.airavata.sharing.model.ResourceTypeEnum;
-import org.apache.airavata.sharing.model.StateEnum;
-import org.apache.airavata.sharing.repository.ResourceRepository;
+import org.apache.airavata.research.repository.ResourceRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
