+++ airavata-api/research-service/src/main/java/org/apache/airavata/research/service/ResearchResourceService.java
@@ -27,18 +27,18 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.airavata.config.UserContext;
 import org.apache.airavata.model.user.proto.UserProfile;
+import org.apache.airavata.research.model.PrivacyEnum;
 import org.apache.airavata.research.model.RepositoryResourceEntity;
+import org.apache.airavata.research.model.ResourceEntity;
+import org.apache.airavata.research.model.ResourceStarEntity;
+import org.apache.airavata.research.model.ResourceTypeEnum;
+import org.apache.airavata.research.model.StateEnum;
+import org.apache.airavata.research.model.StatusEnum;
+import org.apache.airavata.research.model.TagEntity;
 import org.apache.airavata.research.repository.ResearchProjectRepository;
-import org.apache.airavata.sharing.model.PrivacyEnum;
-import org.apache.airavata.sharing.model.ResourceEntity;
-import org.apache.airavata.sharing.model.ResourceStarEntity;
-import org.apache.airavata.sharing.model.ResourceTypeEnum;
-import org.apache.airavata.sharing.model.StateEnum;
-import org.apache.airavata.sharing.model.StatusEnum;
-import org.apache.airavata.sharing.model.TagEntity;
-import org.apache.airavata.sharing.repository.ResourceRepository;
-import org.apache.airavata.sharing.repository.ResourceStarRepository;
-import org.apache.airavata.sharing.repository.TagRepository;
+import org.apache.airavata.research.repository.ResourceRepository;
+import org.apache.airavata.research.repository.ResourceStarRepository;
+import org.apache.airavata.research.repository.TagRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -127,9 +127,9 @@ public class ResearchResourceService {
         resource.setName(name);
         resource.setDescription(description);
         resource.setAuthors(authors.stream().map(String::toLowerCase).collect(Collectors.toSet()));
-        Set<org.apache.airavata.sharing.model.TagEntity> tagsSet = new HashSet<>();
+        Set<org.apache.airavata.research.model.TagEntity> tagsSet = new HashSet<>();
         for (String tag : tags) {
-            org.apache.airavata.sharing.model.TagEntity t = new org.apache.airavata.sharing.model.TagEntity();
+            org.apache.airavata.research.model.TagEntity t = new org.apache.airavata.research.model.TagEntity();
             t.setValue(tag);
             tagsSet.add(t);
         }
